requirements.txt
Django==4.2.30
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
openai==1.59.8
anthropic==0.42.0
tiktoken==0.8.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.11
dj-database-url==2.1.0
ijson==3.3.0
certifi==2025.1.31
